Put all components in separate files
